fix: skip Temporal cleanup in preview environment

+++ lib/kube/clean.sh
@@ -30,10 +30,15 @@ fi
 
 if [ -d "$kube_env_dir" ]; then
     for file in "$kube_env_dir"/*; do
+        if [[ "$KUBE_ENV" == "preview" && "$file" == *temporal-deployment.yaml ]]; then
+            echo "clean :: skipping temporal cleanup in preview env - $file"
+            continue
+        fi
         echo "clean :: cleaning from env config - $kube_env_dir/$file"
         envsubst <"$file" | kubectl delete --ignore-not-found=true -f -
     done
 fi
+
 
 # deployment post clean hook
 if [ -f "$kube_post_clean_script" ]; then
